refactor text-diamond-bounding-polygon test
